+++ src/lib/components/forms/reveal-secret-form.svelte
@@ -9,12 +9,12 @@
 
 	import { dev } from '$app/environment';
 	import { createDownloadLinkAndClick, sendMessageToServiceWorker } from '$lib/client/utils';
+	import { decryptString } from '$lib/client/web-crypto';
 	import * as Form from '$lib/components/ui/form';
 	import { type FileMeta, type FileReference, handleFileChunksDownload } from '$lib/file-transfer';
 	import { formatBytes } from '$lib/i18n';
 	import * as m from '$lib/paraglide/messages.js';
 	import { type RevealSecretFormSchema, revealSecretFormSchema } from '$lib/validators/formSchemas';
-	import { decryptString } from '$lib/web-crypto';
 
 	import Password from '../form-fields/password.svelte';
 	import Typewriter from '../helpers/typewriter.svelte';
