--- src/lib/components/ui/usps/index.ts
@@ -1,1 +1,0 @@
-export { default } from './usps.svelte';
