+++ src/routes/(default)/s/+page.svelte
@@ -2,14 +2,15 @@
 	import { onMount } from 'svelte';
 
 	import { api } from '$lib/api';
+	import { sha256Hash } from '$lib/client/web-crypto';
 	import RevealSecretForm from '$lib/components/forms/reveal-secret-form.svelte';
 	import Page from '$lib/components/layout/page/page.svelte';
 	import Alert from '$lib/components/ui/alert/alert.svelte';
 	import { Spinner } from '$lib/components/ui/spinner';
 	import * as m from '$lib/paraglide/messages.js';
-	import { sha256Hash } from '$lib/web-crypto';
 
 	import type { PageData } from './$types';
+
 	let isLoading = $state(true);
 	let error: string = $state('');
